HBX-1703: Move class 'org.hibernate.tool.hbm2x.HibernateConfigurationExporter' to 'org.hibernate.tool.internal.export.cfg.HibernateConfigurationExporter'
